requirements.txt
cached-property==1.5.2
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
importlib-metadata==4.12.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
mypy-extensions==0.4.3
packaging==21.3
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.41
sqlalchemy-orm==1.2.3
typing-inspect==0.8.0
typing_extensions==4.3.0
waitress==2.1.2
Werkzeug==2.2.2
zipp==3.8.1